phase 3 (3/6 fixup): move tokenizers stub to package __init__
llm.py's tokenizers stub only fired when llm was imported — but
the runner / tools modules / manual inspections can hit litellm
without going through llm first. Moving the stub to
dportsv3/agent/__init__.py makes it run as soon as any module
under the package is imported.

This unblocks invocations like:
  python -c "import dportsv3.agent; import litellm; ..."

without needing to pre-import llm.
